requirements.txt
playwright==1.39.0
psycopg2-binary==2.9.7
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.26.3
lxml==4.9.3
python-dotenv==1.0.1
tqdm==4.66.1
